requirements.txt
joblib
kivy
matplotlib
numexpr
numpy
pafy
pandas
python-nmap
python-vlc
scipy
sklearn
youtube-dl==2020.12.2